Avoid recompiling previous tests
This is done by:
- Storing compiled executable into a new "Build" folder
- When running a new tests, comparing compiling options to those used
in previous tests; it they match, the previous executable is retrieved
from the Build folder.

Note: this assumes that each executable name is unique ; i.e. that
two sets of compiling options cannot produce the same executable name.

In order to obtain the string that contains compiling options, part
of the function `build_c` was extracted as a separate function