Fail loudly if new benchmark has no output.
When --make_benchmarks is on, a misconfigured / buggy test program might
fail to produce any output. If this happens, the regression testing
script will continue executing and eventually crash when it can't find
the .status file that it is expecting to be there. This produces an
unintuitive backtrace that does not make clear the root cause of the
problem.

This PR fixes the above by failing with a clear error message as soon as
it can be determined that output from the test program cannot be found.